Refactor - move GetFilesOfInterest type and use the appropriate define
